Add text based game and image manipulation
The "Game" folder is for a text based game developed
while working through the Learn Python In One Day
And Learn It Well book.

The image manipulation file was also from that book,
although I played with a Doctor Who file rather than the
supplied one. It's simple but demonstrated useful
concepts.